requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
PyMySQL==1.0.2
bcrypt==3.2.0
Werkzeug==2.0.1
pytest==6.2.5
black==21.7b0
flake8==3.9.2
mypy==0.910 